Changed Mileage from double to BigDecimal

+++ src/main/java/ru/fisher/VehiclePark/dto/MileageReportDTO.java
@@ -3,6 +3,7 @@ package ru.fisher.VehiclePark.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 @Getter
@@ -12,5 +13,5 @@ public class MileageReportDTO {
     private String period;
     private String startDate;
     private String endDate;
-    private Map<String, Double> results; // Пары "время-значение"
+    private Map<String, BigDecimal> results; // Пары "время-значение"
 }
